requirements.txt
# Core LLM & AI
anthropic>=0.25.0
openai>=1.30.0
tiktoken>=0.7.0

# RAG & Vector DB
faiss-cpu>=1.8.0
sentence-transformers>=3.0.0
langchain>=0.2.0
langchain-community>=0.2.0
langchain-anthropic>=0.1.15

# Web Scraping & APIs
httpx>=0.27.0
aiohttp>=3.9.0
beautifulsoup4>=4.12.0
feedparser>=6.0.11
playwright>=1.44.0
newspaper3k>=0.2.8
lxml>=5.2.0

# API Framework
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0
pydantic-settings>=2.3.0

# Async & Concurrency
asyncio-throttle>=1.0.2
aiofiles>=23.2.1
tenacity>=8.3.0

# Data Processing
numpy>=1.26.0
pandas>=2.2.0
python-dateutil>=2.9.0

# Caching & Storage
redis>=5.0.0
diskcache>=5.6.3

# Monitoring & Logging
structlog>=24.1.0
prometheus-client>=0.20.0
rich>=13.7.0

# Configuration
python-dotenv>=1.0.1
pyyaml>=6.0.1

# Testing
pytest>=8.2.0
pytest-asyncio>=0.23.0
pytest-cov>=5.0.0
httpx[testing]>=0.27.0

# Dev tools
black>=24.4.0
ruff>=0.4.0
mypy>=1.10.0
pre-commit>=3.7.0
